edge-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "can-telemetry-edge-agent"
version = "1.0.0"
description = "CAN bus data capture and upload agent for telemetry platform"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "CAN Telemetry Team", email = "team@example.com"}
]
dependencies = [
    "python-can>=4.3.1",
    "cantools>=39.4.5",
    "pyarrow>=15.0.0",
    "boto3>=1.34.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "moto[s3]>=5.0.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
